pyproject.toml
[project]
name = "autonomous-lakehouse"
version = "0.1.0"
description = "Autonomous Lakehouse Operations Platform — AI control plane for self-healing data pipelines"
readme = "README.md"
license = "AGPL-3.0-only"
license-files = ["LICENSE"]
requires-python = ">=3.12"
dependencies = [
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "slowapi>=0.1.9",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Configuration
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    # Observability
    "structlog>=24.4.0",
    "prometheus-client>=0.21.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    # Task queue
    "arq>=0.26.1",
    # Authentication
    "PyJWT[crypto]>=2.9.0",
    # Utilities
    "httpx>=0.28.0",
    "python-ulid>=3.0.0",
]

[project.optional-dependencies]
llm = [
    "anthropic>=0.40.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "anthropic>=0.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lakehouse"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "S",    # flake8-bandit (security)
    "T20",  # flake8-print
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
]
ignore = [
    "S101",  # allow assert in tests
    "B008",  # Depends() in function defaults is standard FastAPI pattern
]

[tool.ruff.lint.per-file-ignores]
"src/lakehouse/adapters/mock.py" = ["S311"]  # pseudo-random is fine for mock data
"src/lakehouse/worker/tasks.py" = ["RUF012"]  # arq WorkerSettings requires class-level mutables

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["error"]
markers = [
    "integration: marks tests as integration tests (require real cloud credentials)",
    "timeout: marks tests with a timeout value",
]

[tool.coverage.run]
source = ["src/lakehouse"]
omit = ["*/migrations/*"]
